History page - Moved History interface to core package; inserted run status image to historyRow in the beginning; added link to download script from history
