+++ bundle/src/main/java/com/cognifide/cq/cqsm/core/models/ExecutionHistoryModel.java
@@ -20,7 +20,7 @@
 package com.cognifide.cq.cqsm.core.models;
 
 import com.cognifide.cq.cqsm.api.history.Entry;
-import com.cognifide.cq.cqsm.api.history.History;
+import com.cognifide.cq.cqsm.core.history.History;
 
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.collections.Predicate;
