+++ bundle/src/main/java/com/cognifide/cq/cqsm/core/scripts/listeners/HistoryReplicateListener.java
@@ -21,7 +21,7 @@ package com.cognifide.cq.cqsm.core.scripts.listeners;
 
 import com.cognifide.cq.cqsm.api.executors.Mode;
 import com.cognifide.cq.cqsm.api.history.Entry;
-import com.cognifide.cq.cqsm.api.history.History;
+import com.cognifide.cq.cqsm.core.history.History;
 import com.cognifide.cq.cqsm.api.logger.Progress;
 import com.cognifide.cq.cqsm.api.scripts.Event;
 import com.cognifide.cq.cqsm.api.scripts.EventListener;
