+++ content/src/main/content/jcr_root/apps/apm/clientlibs/history/css/history.less
@@ -17,16 +17,12 @@
  * limitations under the License.
  * =========================LICENSE_END==================================
  */
-.dryRunFailed, .runFailed {
-  min-height: 20px;
-  text-align: center;
-  background: url(../img/icons/error.gif) no-repeat center center;
+img.dryRunFailed, img.runFailed {
+  content: url(../img/icons/error.gif);
 }
 
-.dryRunSuccess, .runSuccess {
-  min-height: 20px;
-  text-align: center;
-  background: url(../img/icons/check.gif) no-repeat center center;
+img.dryRunSuccess, img.runSuccess {
+  content: url(../img/icons/check.gif);
 }
 .summary {
   background: url(../img/icons/summary.png) no-repeat center left;
